% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_selector.R
\name{compute_bic}
\alias{compute_bic}
\title{Compute BIC for Fitted Mixture Model}
\usage{
compute_bic(model, X, y)
}
\arguments{
\item{model}{A fitted mixture model (from fit_mixture_regression)}

\item{X}{Predictor matrix}

\item{y}{Response vector}
}
\value{
Numeric BIC value
}
\description{
Compute BIC for Fitted Mixture Model
}
